backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.8.3
anthropic==0.39.0
supabase==2.9.1
asyncpg==0.30.0
sqlalchemy==2.0.36
alembic==1.13.3
celery[redis]==5.4.0
redis==5.1.1
boto3==1.35.36
sentry-sdk[fastapi]==2.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyYAML==6.0.2
